TremorDetectAI/requirements.txt
pyserial==3.5
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2 